[obey-campaign:8deed8b4] test(clone,sync): containerize execution via build tag instead of relocating
R4 asked for the filesystem-mutating internal/clone and internal/sync suites to
move into the containerized lane. Taken literally that could not be done: 61 of
these tests mutate the filesystem and most assert on UNEXPORTED package
internals -- listSubmodules, RunPreflight, updateSubmodules, the Cloner git*
methods, reverseSyncLocalURLs, verifyArtifacts. tests/integration is package
integration, so relocating them deletes those assertions rather than moving
them, and several have no CLI expression at all: no camp command prints the
slice listSubmodules returns.

The containerization rule exists so a test that creates or git-inits real files
cannot do it on a developer's machine; location in tests/integration was only
ever the proxy for that. So this containerizes execution instead. FS-mutating
tests move to *_container_test.go files tagged //go:build container_fs, which
the host never compiles, and the lane cross-compiles each package's test binary
for linux/musl, copies it into a pooled container and runs it there. No test
body rewritten, nothing exported, no assertion lost.

136 original tests = 75 host + 61 tagged, verified per-file against HEAD. The
container reports 101 and 35 top-level tests actually executed, and the harness
asserts a per-package minimum precisely because a tag that silently excluded
everything from everywhere would be the worst outcome -- which it caught: the
runner was first named containerfs.go, so Go never collected it and the lane
reported 'no tests to run' as a pass.

The ratchet becomes a location-or-tag rule, checked mechanically in a file's
first three lines, keeping its fail-closed posture. Its textual grep also turns
out to under-detect: clone_test.go and gitmodules_test.go mutate through shared
helpers rather than a literal exec.Command("git"), so they were never
allowlisted and never flagged. Scope is the package, so both were migrated.

Recorded as D007, with the ratchet change flagged for veto. The cmdutil and
internal/git entries stay out of scope with their intent.
