+++ justfile
@@ -76,13 +76,36 @@ lint:
 lint-no-host-fs-tests:
     #!/usr/bin/env sh
     set -eu
-    # Allowlist: existing host-FS test violators tracked under CW0003-tests-01
-    # and adjacent legacy patterns. Goal: drive this list to empty by migrating
-    # tests to tests/integration/ + the container harness. The rule blocks any
-    # NEW addition beyond this set.
-    allowlist="./internal/git/remote_test.go ./internal/git/commit_test.go ./internal/git/info_exclude_test.go ./pkg/commitkit/commitkit_test.go ./cmd/camp/project/remote/remote_test.go ./tools/release-notes/main_test.go ./internal/doctor/checks/orphan_test.go ./internal/doctor/checks/head_test.go ./internal/doctor/checks/url_test.go ./internal/project/add_test.go ./internal/project/resolve_test.go ./internal/project/list_test.go ./internal/attach/attach_test.go ./internal/leverage/backfill_test.go ./internal/leverage/authors_test.go ./internal/leverage/projects_test.go ./internal/leverage/blame_cache_test.go ./internal/leverage/sampler_test.go ./internal/leverage/config_test.go ./internal/clone/git_test.go ./internal/quest/autocommit_integration_test.go ./internal/sync/sync_test.go ./internal/sync/preflight_test.go ./internal/scaffold/init_behavior_test.go ./internal/git/commit/commit_test.go ./internal/git/executor_test.go ./internal/git/submodule_test.go ./internal/git/submodule_list_test.go ./internal/git/branches_test.go ./internal/git/submodule_orphan_test.go ./internal/git/resolve_test.go"
-    hits=$(find . -name '*_test.go' -not -path './tests/integration/*' -not -path './vendor/*' -print0 2>/dev/null | \
+    # A test that mutates real filesystem state is compliant when it cannot run
+    # on a developer's machine. There are two ways to achieve that, and the rule
+    # accepts both (decision D007):
+    #
+    #   1. it lives under tests/integration/ and drives the camp binary, or
+    #   2. it carries the container_fs build tag, so a plain `just test` does
+    #      not compile it, and the integration lane cross-compiles it and runs
+    #      it inside a pooled container (tests/integration/containerfs_test.go).
+    #
+    # Form 2 exists because the internal/clone and internal/sync suites assert
+    # on unexported package internals; relocating them into package integration
+    # would have deleted those assertions rather than moved them.
+    #
+    # Allowlist: legacy host-FS violators tracked under CW0003-tests-01 and
+    # adjacent patterns, still to migrate. The rule blocks any NEW addition
+    # beyond this set. Goal: drive this list to empty.
+    allowlist="./internal/git/remote_test.go ./internal/git/commit_test.go ./internal/git/info_exclude_test.go ./pkg/commitkit/commitkit_test.go ./cmd/camp/project/remote/remote_test.go ./tools/release-notes/main_test.go ./internal/doctor/checks/orphan_test.go ./internal/doctor/checks/head_test.go ./internal/doctor/checks/url_test.go ./internal/project/add_test.go ./internal/project/resolve_test.go ./internal/project/list_test.go ./internal/attach/attach_test.go ./internal/leverage/backfill_test.go ./internal/leverage/authors_test.go ./internal/leverage/projects_test.go ./internal/leverage/blame_cache_test.go ./internal/leverage/sampler_test.go ./internal/leverage/config_test.go ./internal/quest/autocommit_integration_test.go ./internal/scaffold/init_behavior_test.go ./internal/git/commit/commit_test.go ./internal/git/executor_test.go ./internal/git/submodule_test.go ./internal/git/submodule_list_test.go ./internal/git/branches_test.go ./internal/git/submodule_orphan_test.go ./internal/git/resolve_test.go"
+    # Candidates: FS-mutating tests outside tests/integration/. Files carrying
+    # the container_fs tag are excluded here rather than allowlisted, because
+    # they are compliant by construction: the tag keeps them off the host and
+    # the lane runs them in a container.
+    candidates=$(find . -name '*_test.go' -not -path './tests/integration/*' -not -path './vendor/*' -print0 2>/dev/null | \
         xargs -0 grep -lE 'exec\.Command\("git"|exec\.CommandContext\(.*"git"' 2>/dev/null || true)
+    hits=""
+    for c in $candidates; do
+        if head -3 "$c" | grep -q '^//go:build container_fs'; then
+            continue
+        fi
+        hits="$hits $c"
+    done
     new_violators=""
     for hit in $hits; do
         case " $allowlist " in
@@ -94,8 +117,11 @@ lint-no-host-fs-tests:
         echo "FAIL: NEW host-side git exec.Command in _test.go outside tests/integration/:"
         for v in $new_violators; do echo "  $v"; done
         echo ""
-        echo "Migrate to tests/integration/ via GetSharedContainer + RunCampInDir,"
-        echo "or (if intentional pure-logic test) add to the allowlist in justfile."
+        echo "Fix by either:"
+        echo "  - moving it to tests/integration/ (GetSharedContainer + RunCampInDir), or"
+        echo "  - tagging the file //go:build container_fs and registering its package"
+        echo "    in containerFSPackages (tests/integration/containerfs_test.go), or"
+        echo "  - (if it is genuinely pure-logic) adding it to the allowlist in justfile."
         exit 1
     fi
     echo "lint-no-host-fs-tests: clean (no NEW violators; $(echo $hits | wc -w | tr -d ' ') legacy files on allowlist)"
